% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimresp.R
\name{baseline_one_trace}
\alias{baseline_one_trace}
\title{Baseline-correct one trace}
\usage{
baseline_one_trace(df, baseline_window_ms = c(300, 400))
}
\arguments{
\item{df}{Data frame with columns \code{time_ms} and \code{value}.}

\item{baseline_window_ms}{Numeric length-2 vector giving the baseline window
in milliseconds.}
}
\value{
A data frame with baseline-corrected \code{value}.
}
\description{
Baseline-correct one trace
}
\keyword{internal}
