% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrad_wl_amp.R
\name{get_zeus_irrad_calibration}
\alias{get_zeus_irrad_calibration}
\title{Get a registered irradiance calibration table}
\usage{
get_zeus_irrad_calibration(name)
}
\arguments{
\item{name}{Calibration identifier.}
}
\value{
A calibration tibble.
}
\description{
Get a registered irradiance calibration table
}
