% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrad_wl_amp.R
\name{nd_to_log_irradiance}
\alias{nd_to_log_irradiance}
\title{Join log irradiance values to stimulus metadata}
\usage{
nd_to_log_irradiance(meta, calib)
}
\arguments{
\item{meta}{Data frame containing \code{wavelength} and \code{stim_nd}.}

\item{calib}{Calibration table containing \code{wavelength}, \code{stim_nd}, and
\code{log_hv}.}
}
\value{
Input metadata with \code{log_hv} added.
}
\description{
Join log irradiance values to stimulus metadata
}
