% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERG_Import.R
\name{read_abf_raw}
\alias{read_abf_raw}
\title{Import a raw ABF file into a ZEUS raw object}
\usage{
read_abf_raw(path, ...)
}
\arguments{
\item{path}{Path to a \code{.abf} file.}

\item{...}{Additional arguments passed to \code{\link[readABF:readABF]{readABF::readABF()}}.}
}
\value{
An object of class \code{zeus_abf_raw}.
}
\description{
Reads an Axon Binary Format (\code{.abf}) file using \code{\link[readABF:readABF]{readABF::readABF()}} and
wraps the result in a \code{zeus_abf_raw} object for downstream ZEUS processing.
}
