% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimresp.R
\name{summarize_stimresp_qc}
\alias{summarize_stimresp_qc}
\title{Summarize StimResp noisy-replicate exclusion}
\usage{
summarize_stimresp_qc(x)
}
\arguments{
\item{x}{A ZEUS object containing \code{stimresp_qc}.}
}
\value{
A list with:
\describe{
\item{overall}{Overall counts of retained and excluded replicates.}
\item{by_stim}{Retained and excluded replicate counts by stimulus.}
\item{dropped}{A table of excluded replicates.}
}
}
\description{
Summarize StimResp noisy-replicate exclusion
}
