% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_exporter.R
\name{zeus_prepare_csv_export}
\alias{zeus_prepare_csv_export}
\title{Prepare one ZEUS component for CSV export}
\usage{
zeus_prepare_csv_export(x, name)
}
\arguments{
\item{x}{Object to export.}

\item{name}{Component name.}
}
\value{
A data frame.
}
\description{
Prepare one ZEUS component for CSV export
}
\keyword{internal}
